refactor(color): delete unused codes & rename comonents
